+++ dependencies/pom.xml
@@ -39,6 +39,13 @@
 				<type>pom</type>
 				<scope>import</scope>
 			</dependency>
+			<dependency>
+				<groupId>org.eclipse.dirigible.tests</groupId>
+				<artifactId>dirigible-tests-framework</artifactId>
+				<version>10.0.0-SNAPSHOT</version>
+				<type>pom</type>
+				<scope>import</scope>
+			</dependency>
 
 			<dependency>
 				<groupId>org.apache.activemq</groupId>
