+++ tests/integration-tests/src/test/java/org/eclipse/dirigible/integration/tests/api/java/messaging/MessagesHolder.java
