+++ tests/integration-tests/src/test/java/org/eclipse/dirigible/integration/tests/services/integrations/CamelPlatformHttpTest.java
