+++ tests/integration-tests/src/test/java/org/eclipse/dirigible/integration/tests/ui/Dirigible.java
