+++ tests/integration-tests/src/test/resources/META-INF/dirigible/camel_test/calledFromCamel.mjs
