+++ tests/integration-tests/src/test/resources/META-INF/dirigible/integration-tests-project/test-queue-listener.listener
