+++ tests/integration-tests/src/test/resources/META-INF/dirigible/modules-tests/src/services/integrations/dirigible-js-to-camel-route-test.js
