+++ tests/integration-tests/src/test/resources/META-INF/dirigible/modules-tests/tsconfig.json
