+++ tests/pom.xml
@@ -1,153 +1,27 @@
+<?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>org.eclipse.dirigible</groupId>
+        <artifactId>dirigible-parent</artifactId>
+        <version>10.0.0-SNAPSHOT</version>
+    </parent>
 
-	<parent>
-		<groupId>org.eclipse.dirigible</groupId>
-		<artifactId>dirigible-parent</artifactId>
-		<version>10.0.0-SNAPSHOT</version>
-		<relativePath>../pom.xml</relativePath>
-	</parent>
+    <name>Tests - Parent</name>
+    <modules>
+        <module>tests-framework</module>
+    </modules>
+    <groupId>org.eclipse.dirigible.tests</groupId>
+    <artifactId>tests</artifactId>
+    <packaging>pom</packaging>
 
-	<properties>
-		<license.header.location>../licensing-header.txt</license.header.location>
-		<parent.pom.folder>../</parent.pom.folder>
-		<skipITs>true</skipITs>
-		<selenide.version>7.0.4</selenide.version>
-	</properties>
-
-	<name>Integration Tests</name>
-	<artifactId>dirigible-integration-tests</artifactId>
-	<packaging>jar</packaging>
-
-	<dependencies>
-		<dependency>
-			<groupId>org.eclipse.dirigible</groupId>
-			<artifactId>dirigible-application</artifactId>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-test</artifactId>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.security</groupId>
-			<artifactId>spring-security-test</artifactId>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>com.codeborne</groupId>
-			<artifactId>selenide</artifactId>
-			<version>${selenide.version}</version>
-			<scope>test</scope>
-		</dependency>
-	</dependencies>
-
-	<dependencyManagement>
-		<dependencies>
-			<dependency>
-				<groupId>org.eclipse.dirigible</groupId>
-				<artifactId>dirigible-application</artifactId>
-				<version>${project.version}</version>
-				<type>pom</type>
-				<scope>import</scope>
-			</dependency>
-		</dependencies>
-	</dependencyManagement>
-
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-surefire-plugin</artifactId>
-				<configuration>
-					<skipTests>true</skipTests>
-				</configuration>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-failsafe-plugin</artifactId>
-				<configuration>
-					<skipITs>${skipITs}</skipITs>
-				</configuration>
-				<executions>
-					<execution>
-						<goals>
-							<goal>integration-test</goal>
-							<goal>verify</goal>
-						</goals>
-					</execution>
-				</executions>
-			</plugin>
-			<plugin>
-				<groupId>org.codehaus.mojo</groupId>
-				<artifactId>exec-maven-plugin</artifactId>
-				<configuration>
-					<skip>${skipITs}</skip>
-				</configuration>
-				<executions>
-					<execution>
-						<id>build-ts-test-sources</id>
-						<phase>generate-test-sources</phase>
-						<goals>
-							<goal>exec</goal>
-						</goals>
-						<configuration>
-							<executable>
-								${basedir}/src/test/resources/META-INF/dirigible/modules-tests/build-source${script.extension}
-							</executable>
-							<workingDirectory>
-								src/test/resources/META-INF/dirigible/modules-tests
-							</workingDirectory>
-						</configuration>
-					</execution>
-				</executions>
-			</plugin>
-		</plugins>
-	</build>
-
-	<profiles>
-		<profile>
-			<id>integration-tests</id>
-			<properties>
-				<skipITs>false</skipITs>
-			</properties>
-		</profile>
-		<profile>
-			<id>Windows 11</id>
-			<activation>
-				<os>
-					<family>dos</family>
-				</os>
-			</activation>
-			<properties>
-				<script.extension>.bat</script.extension>
-			</properties>
-		</profile>
-		<profile>
-			<id>Windows</id>
-			<activation>
-				<os>
-					<family>Windows</family>
-				</os>
-			</activation>
-			<properties>
-				<script.extension>.bat</script.extension>
-			</properties>
-		</profile>
-		<profile>
-			<id>unix</id>
-			<activation>
-				<os>
-					<family>unix</family>
-				</os>
-			</activation>
-			<properties>
-				<script.extension>.sh</script.extension>
-			</properties>
-		</profile>
-	</profiles>
+    <properties>
+        <license.header.location>../licensing-header.txt</license.header.location>
+        <parent.pom.folder>../</parent.pom.folder>
+        <skipITs>true</skipITs>
+        <selenide.version>7.0.4</selenide.version>
+    </properties>
 
 </project>
