+++ tests/tests-framework/src/main/java/org/eclipse/dirigible/tests/framework/Browser.java
@@ -8,7 +8,7 @@
  * SPDX-FileCopyrightText: 2023 SAP SE or an SAP affiliate company and Eclipse Dirigible
  * contributors SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.dirigible.integration.tests.ui.framework;
+package org.eclipse.dirigible.tests.framework;
 
 import com.codeborne.selenide.SelenideElement;
