+++ tests/tests-framework/src/main/java/org/eclipse/dirigible/tests/framework/HtmlAttribute.java
@@ -8,22 +8,19 @@
  * SPDX-FileCopyrightText: 2023 SAP SE or an SAP affiliate company and Eclipse Dirigible
  * contributors SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.dirigible.integration.tests.ui.framework;
+package org.eclipse.dirigible.tests.framework;
 
-public enum HtmlElementType {
-    BUTTON("button"), //
-    INPUT("input"), //
-    ANCHOR("a"), //
-    HEADER5("h5"), //
-    TITLE("title");
+public enum HtmlAttribute {
+    ID("id"), //
+    TYPE("type");
 
-    private final String type;
+    private final String attribute;
 
-    HtmlElementType(String type) {
-        this.type = type;
+    HtmlAttribute(String attribute) {
+        this.attribute = attribute;
     }
 
-    public String getType() {
-        return type;
+    public String getAttribute() {
+        return attribute;
     }
 }
